run rsync up to 3 more times to compare checksums
I had to manually create the target cardname dir instead of relying on
rsync to do so. Indeed, otherwise the log gets messed up as the created
dir is added before the file name. This was a problem when using the log
for select the files to send again.

Also used the long name for rsync's option this way their meaning is
more explicit.

Used the option `--info=PROGRESS2` in first rsync run to get overall
progress instead of per file's one.

Closes #19
Closes #23